Updated queryselector paths for WPM and Accuracy.
I got in contact with the developer of the website and they were kind enough to add some specific hidden HTML paths for me to get WPM and accuracy so it would be accurate 100% of the time. I have asked them to get things like time and race place which I will update once they add it to the website. This is because I realized that the presence wasn't always displaying the most accurate information because of the order in which users join the race. It is very rare that the information be inaccurate but I am still implementing these features so I know it will be accurate 100% of the time.