Added Typer.io Presence (#3956)
* Added Typer.io Presence

* Hopefully fixed all Codacy Static Code Analysis issues and rechecked everything works after edits.

* Fixed last few Codacy Static Code Analysis problems.

* Hopefully fixed last Codacy problem.

* Fixed more minor bracket spacing issues with Codacy (hopefully last fix)

* Fixed the rest of the bracket stuff, this should be the last.

* Curly brace 😳

* Update keywords.

Co-authored-by: ririxidev <mail@ririxi.dev>

* Removed unneeded if statement and ran prettier on the presence file.

* Extra pretified the code 👍

* Fixed comma problems caused by pretify.

* Correctly pretified the file.

* Updated queryselector paths for WPM and Accuracy.

I got in contact with the developer of the website and they were kind enough to add some specific hidden HTML paths for me to get WPM and accuracy so it would be accurate 100% of the time. I have asked them to get things like time and race place which I will update once they add it to the website. This is because I realized that the presence wasn't always displaying the most accurate information because of the order in which users join the race. It is very rare that the information be inaccurate but I am still implementing these features so I know it will be accurate 100% of the time.

* Ran prettier on the most recent code.

* Added more hidden elements and removed no longer needed replace arguments.

* Ran prettier on most recent pull request.

* Changed "did not place" to "did not finish"

* Removed .split("Jump to"[0]

It was no longer needed after my previous updates.

* Ran prettier on most recent update.

* Updated "viewing a users profile"

Instead of having the username be a separate state, it is now apart of the details area.

Co-authored-by: ririxidev <mail@ririxi.dev>

Co-authored-by: ririxidev <mail@ririxi.dev>