feat(Stremio): add Stremio 5 & rewrite presence (#7478)
* refactor(Stremio): improve control flow

Replace nested if/else statements with switch statements. This makes the code easier to read and maintain going forward.

* feat(Stremio): add Stremio 5

Add support for stremio 5 (a.k.a Stremio web).

* refactor(Stremio): simplify player page handling

There's no need to parse the timestamps from HTML elements because the video element, which contains enough information already, is always available on the player page.

* fix(Stremio): replace link to player with meta URL

The button link to the player URL is 1) too long for discord 2) forbidden by the PreMiD developer guidelines. So, we link to the meta details page instead.

* feat(Stremio): show thumbnail for app version 4

* refactor(Stremio): improve handling undefined values

* fix(Stremio): update invalid CSS selectors

* feat(Stremio): add privacy-friendly search

* fix(Stremio): link to correct meta URLs

* style(Stremio): improve details and state wording

* feat(Stremio): show metadata info for channels

* feat(Stremio): add fallback for non-video player

Add a fallback for content being played without native `<video>` element. This is useful for things like channels, where the video is embedded by an iframe to another site (like YouTube).

* refactor(Stremio): use attribute contains selector

This selector withstands class reordering.

* refactor(Stremio): store logo in Assets enum

* feat(Stremio): add small image text

* feat(Stremio): show content type from meta page

* refactor(Stremio): reorder state and details

* feat(Stremio): add optional search queries

Due to the nature of some content on Stremio (e.g. adult content) there should be a setting to enable displaying search queries.

* refactor(Stremio): remove global title variable

* refactor(Stremio): perform thumbnails check once

* refactor(Stremio): move 'find player info' logic to separate function

* fix(Stremio): fix logo not showing

* refactor(Stremio): remove incorrect selectors

These are either wrong or null when the page reloads so there's no point in keeping them.

* chore(Stremio): format code

* fix(Stremio): fix typo

* feat(Stremio): add more buttons

* fix(Stremio): fix incorrect state of loading player

* feat(Stremio): include genre from discover page

* chore(Stremio): format code

* chore(Stremio): fix deepscan issues

* chore(Stremio): swap contributor with author

* chore(Stremio): fix typo

* refactor(Strremio): apply destructuring pattern

* chore(Stremio): capitalize labels