Fixed MangaDex presence and chapter/page display issues
Updated the script to include logic for extracting the chapter number and page number from the MangaDex page when the user is reading a manga chapter. This information is then used to set the presence data accordingly.

Specifically, in the "case 'chapter'" section of the script, added code to extract the chapter number and page number from the page elements. These values are then used to set the presenceData.state property to display the chapter and page information.

Additionally, tested and confirmed the functionality on the Chrome browser running on Windows 11.
https://github.com/PreMiD/Presences/issues/7412

Signed-off-by: Riza <90960705+findkeys@users.noreply.github.com>