fix(Stremio): resolve player state promise before submitting the event (#8540)
* fix(Stremio): resolve player state promise before submitting the event

There's an issue with Chromium-based browsers (such as Chrome and Microsoft Edge) in particular where the received player state is always `null`. On Firefox this issue does not occur. I'm not 100% sure of the exact cause of the discrepancy between the two browsers, but it seems like Chromium has trouble resolving promises that were dispatched by the `CustomEvent` event. So, to address this issue we have to resolve the promise first before dispatching the event.

I also added a check to gracefully handle cases where the `metaItem` field is still `undefined`, which may happen to users with slow network speeds.

* chore(Stremio): bump version